feat: update private model example code
Minor update to the example code for private models. The update is meant
to bring the example code in line with the changes made to the Webiny
documentation (at docs.webiny.com).

The wrapper to get the model "unauthorized" has been removed as well.
This code was not present in the documentation anymore and furthermore
the code that is being called already contains that check. So the check
was redundant anyway.

Helped-by: Pavel Denisjuk <pavel@webiny.com>